Server/fastapi-server/requirements.txt
fastapi
uvicorn
asyncpg
pandas
apscheduler
pyampd
python-dotenv
aiomysql
cryptography
pytz
redis
ta
httpx
requests
python-telegram-bot
discord.js
discord